Completed setup of Ellipse, Quadrilateral and Triangles

+++ src/main/java/objects/curve/Circle.java
@@ -3,9 +3,11 @@ package objects.curve;
 
 import objects.curve.Ellipse;
 
+import java.math.BigDecimal;
+
 public class Circle extends Ellipse {
 
-    public Circle(double radius) {
+    public Circle(BigDecimal radius) {
         super(radius, radius);
     }
